scripts/requirements.txt
aioconsole==0.8.1
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.7.0
asgiref==3.8.1
async-timeout==4.0.3
asyncio-atexit==1.0.1
attrs==24.2.0
autogen-agentchat==0.4.0.dev11
autogen-core==0.4.0.dev11
autogen-ext==0.4.0.dev11
Automat==22.10.0
azure-core==1.32.0
azure-cosmos==4.2.0
azure-storage-blob==12.19.0
azure-identity==1.19.0
azure-keyvault==4.2.0
azure-keyvault-certificates==4.8.0
azure-keyvault-keys==4.9.0
azure-keyvault-secrets==4.8.0
azure-kusto-data==4.4.1
Babel==2.10.3
backoff==2.2.1
bcrypt==4.1.3
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.9.0
build==1.2.1
bump2version==1.0.1
cachetools==5.3.3
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
chroma-hnswlib==0.7.3
chromadb==0.5.0
click==8.1.8
cloud-init==24.4.1
cmarkgfm==0.8.0
colorama==0.4.6
coloredlogs==15.0.1
command-not-found==0.3
configobj==5.0.8
constantly==23.10.4
cryptography==44.0.0
dataclasses-json==0.6.6
dbus-python==1.3.2
Deprecated==1.2.15
dirtyjson==1.0.8
diskcache==5.6.3
distro==1.9.0
distro-info==1.7+build1
dnspython==2.6.1
docker==7.1.0
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.20.1
email_validator==2.1.1
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.4
filelock==3.14.0
FLAML==2.3.2
Flask==3.1.0
Flask-Cors==4.0.1
flatbuffers==24.3.25
frozenlist==1.5.0
fsspec==2024.5.0
git-filter-repo==2.45.0
google-api-core==2.24.2
google-auth==2.29.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.64.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httplib2==0.20.4
httptools==0.6.1
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.23.1
humanfriendly==10.0
hyperlink==21.0.0
idna==3.10
ijson==3.3.0
imbalanced-learn==0.12.3
importlib_metadata==8.4.0
importlib_resources==6.4.0
incremental==22.10.0
isodate==0.6.1
itsdangerous==2.2.0
jaraco.classes==3.2.1
jeepney==0.8.0
Jinja2==3.1.6
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.32
jsonpointer==2.0
jsonref==1.1.0
jsonschema==4.10.3
keyring==24.3.1
kubernetes==29.0.0
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
llama-index==0.10.27
llama-index-agent-openai==0.2.5
llama-index-cli==0.1.12
llama-index-core==0.10.38.post2
llama-index-embeddings-azure-openai==0.1.9
llama-index-embeddings-openai==0.1.7
llama-index-indices-managed-llama-cloud==0.1.6
llama-index-legacy==0.9.48
llama-index-llms-azure-openai==0.1.8
llama-index-llms-openai==0.1.15
llama-index-multi-modal-llms-openai==0.1.6
llama-index-program-openai==0.1.6
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.23
llama-index-readers-llama-parse==0.1.4
llama-index-vector-stores-chroma==0.1.8
llama-parse==0.4.3
llamaindex-py-client==0.1.19
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.21.2
mcp==1.6.0
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
more-itertools==10.2.0
mpmath==1.3.0
msal==1.31.1
msal-extensions==1.2.0
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
netifaces==0.11.0
networkx==3.3
nltk==3.8.1
numpy==1.26.4
oauthlib==3.2.2
olefile==0.46
onnxruntime==1.18.0
openai==1.57.2
opencensus==0.11.4
opencensus-context==0.1.3
opencensus-ext-azure==1.1.14
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-asgi==0.45b0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
opentelemetry-util-http==0.45b0
orjson==3.10.3
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pillow==11.0.0
pkginfo==1.9.6
portalocker==2.10.1
posthog==3.5.0
propcache==0.2.1
proto-plus==1.26.1
protobuf==4.25.5
psutil==7.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyautogen==0.5.0
pycparser==2.22
pycurl==7.45.3
pydantic==2.10.3
pydantic-settings==2.8.1
pydantic_core==2.27.1
Pygments==2.18.0
PyGObject==3.48.2
PyHamcrest==2.1.0
PyJWT==2.10.1
pyOpenSSL==23.2.0
pyparsing==3.1.1
pypdf==4.2.0
PyPika==0.48.9
pyproject_hooks==1.1.0
pyrsistent==0.20.0
pyserial==3.5
python-apt==2.7.7+ubuntu4
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.9
pytz==2025.2
PyYAML==6.0.1
readme-renderer==41.0
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.7.1
roman==3.3
rsa==4.9
scikit-learn==1.5.0
scipy==1.14.0
SecretStorage==3.3.3
service-identity==24.1.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.30
sse-starlette==2.2.1
starlette==0.46.1
striprtf==0.0.26
sympy==1.12
systemd-python==235
tenacity==8.3.0
termcolor==2.5.0
texttable==1.6.7
threadpoolctl==3.5.0
tiktoken==0.8.0
tinycss2==1.2.1
tokenizers==0.19.1
tomli==2.0.1
tqdm==4.67.1
twine==5.0.0
Twisted==24.3.0
typer==0.15.2
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.2
ubuntu-pro-client==8001
ujson==5.10.0
unattended-upgrades==0.1
urllib3==2.2.3
uvicorn==0.29.0
uvloop==0.19.0
wadllib==1.3.6
watchfiles==0.21.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.1.3
wrapt==1.17.0
yarl==1.18.3
zipp==3.21.0
zope.interface==6.1
